chore(routes): delete legacy sandbox.ts duplicate (split files own all logic)
The legacy src/server/routes/sandbox.ts contained complete duplicates of
createK8sRoutes, createNomadRoutes, validateNomadAddress, and
createSandboxRoutes — all already extracted into split files
(sandbox-k8s.ts, sandbox-nomad.ts, sandbox-configs.ts) per AR-023.

router.ts wires the split files; nothing in production imports the
legacy file. The duplicates were dragging combined coverage down by
adding 254 unreachable lines to the denominator.

Migrate the four test files that still imported from the legacy path
to the split files. All 220 sandbox tests still pass with new imports.
